GH-791 Move constants into EnedisRegionConnectorMetadata
