+++ region-connectors/region-connector-fr-enedis/src/main/java/energy/eddie/regionconnector/fr/enedis/EnedisRegionConnectorMetadata.java
@@ -4,11 +4,14 @@ import energy.eddie.api.v0.RegionConnectorMetadata;
 
 import javax.annotation.Nullable;
 import java.time.Period;
+import java.time.ZoneId;
 
 public class EnedisRegionConnectorMetadata implements RegionConnectorMetadata {
     public static final String REGION_CONNECTOR_ID = "fr-enedis";
     public static final Period PERIOD_EARLIEST_START = Period.ofYears(-3);
     public static final Period PERIOD_LATEST_END = Period.ofYears(3);
+    public static final long MAXIMUM_MONTHS_IN_THE_PAST = Math.abs(PERIOD_EARLIEST_START.toTotalMonths());
+    public static final ZoneId ZONE_ID_FR = ZoneId.of("Europe/Paris");
 
     @Nullable
     private static EnedisRegionConnectorMetadata instance = null;
