+++ region-connectors/region-connector-fr-enedis/src/main/java/energy/eddie/regionconnector/fr/enedis/permission/request/EnedisPermissionRequest.java
@@ -13,7 +13,7 @@ import java.time.ZonedDateTime;
 import java.util.Optional;
 import java.util.UUID;
 
-import static energy.eddie.regionconnector.fr.enedis.EnedisRegionConnector.ZONE_ID_FR;
+import static energy.eddie.regionconnector.fr.enedis.EnedisRegionConnectorMetadata.ZONE_ID_FR;
 
 @Entity
 @Table(schema = "fr_enedis", name = "enedis_permission_request")
