+++ region-connectors/region-connector-fr-enedis/src/main/java/energy/eddie/regionconnector/fr/enedis/services/FutureDataService.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDate;
 import java.util.List;
 
-import static energy.eddie.regionconnector.fr.enedis.EnedisRegionConnector.ZONE_ID_FR;
+import static energy.eddie.regionconnector.fr.enedis.EnedisRegionConnectorMetadata.ZONE_ID_FR;
 
 @Service
 public class FutureDataService {
