+++ region-connectors/region-connector-fr-enedis/src/test/java/energy/eddie/regionconnector/fr/enedis/services/FutureDataServiceTest.java
@@ -20,7 +20,7 @@ import java.util.Collections;
 import java.util.Optional;
 import java.util.stream.Stream;
 
-import static energy.eddie.regionconnector.fr.enedis.EnedisRegionConnector.ZONE_ID_FR;
+import static energy.eddie.regionconnector.fr.enedis.EnedisRegionConnectorMetadata.ZONE_ID_FR;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
